Migrate SchemaCoordinateStats
